requirements.txt
flask>=3.1.2
pydantic>=2.0.0
google-genai==1.1.0
rotagent==0.1.0
python-dotenv==1.1.0